feat: move json output responsibility from respect-core to cli
