+++ packages/core/src/__tests__/logger.test.ts
@@ -1,13 +1,15 @@
 import { MockInstance } from 'vitest';
 
 import * as colorette from 'colorette';
-import { logger, colorize } from '../logger.js';
+import { logger, colorize, RESET_ESCAPE_CODE_IN_TERMINAL } from '../logger.js';
 
 describe('Logger in nodejs', () => {
   let spyingStderr: MockInstance;
+  let spyingStdout: MockInstance;
 
   beforeEach(() => {
     spyingStderr = vi.spyOn(process.stderr, 'write').mockImplementation(() => true);
+    spyingStdout = vi.spyOn(process.stdout, 'write').mockImplementation(() => true);
   });
 
   it('should call "process.stderr.write" for error severity', () => {
@@ -29,6 +31,25 @@ describe('Logger in nodejs', () => {
 
     expect(spyingStderr).toBeCalledTimes(1);
     expect(spyingStderr).toBeCalledWith('info');
+  });
+
+  it('should call "process.stdout.write" for output printNewLine', () => {
+    logger.printNewLine();
+
+    expect(spyingStdout).toBeCalledTimes(1);
+    expect(spyingStdout).toBeCalledWith(`${RESET_ESCAPE_CODE_IN_TERMINAL}\n`);
+  });
+
+  it('should call "process.stdout.write" for output printSeparator', () => {
+    logger.printSeparator('separator');
+
+    expect(spyingStdout).toBeCalledTimes(1);
+    expect(spyingStdout).toBeCalledWith(expect.stringContaining(colorize.gray('separator')));
+  });
+
+  it('should call output with indent', () => {
+    const indent = logger.indent('indent', 2);
+    expect(indent).toEqual('\xa0\xa0indent');
   });
 });
