+++ packages/core/src/logger.ts
@@ -2,6 +2,8 @@ import * as colorette from 'colorette';
 import { isBrowser } from './env.js';
 import { identity } from './utils.js';
 
+export const RESET_ESCAPE_CODE_IN_TERMINAL = process.env.NO_COLOR ? '' : '\x1B[0m';
+
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore this works but some types are not working
 export const colorOptions = colorette.options;
@@ -31,6 +33,30 @@ class Logger {
 
   output(str: string) {
     return isBrowser ? console.info(str) : process.stdout.write(str);
+  }
+
+  printNewLine() {
+    return isBrowser
+      ? console.info('\n')
+      : process.stdout.write(`${RESET_ESCAPE_CODE_IN_TERMINAL}\n`);
+  }
+
+  printSeparator(separator: string) {
+    const windowWidth = process.stdout.columns || 80;
+    const coloredSeparator = isBrowser ? separator : colorize.gray(separator);
+    const separatorLine = coloredSeparator
+      .repeat(Math.ceil(windowWidth / separator.length))
+      .slice(0, windowWidth);
+
+    return isBrowser ? console.info(separatorLine) : process.stdout.write(separatorLine);
+  }
+
+  indent(str: string, level: number) {
+    const indentChar = isBrowser ? '  ' : '\xa0';
+    return str
+      .split('\n')
+      .map((line) => indentChar.repeat(level) + line)
+      .join('\n');
   }
 }
