+++ packages/respect-core/src/index.ts
@@ -1,4 +1,6 @@
 export { generateArazzo, handleRun } from './handlers/index.js';
 export type { GenerateArazzoOptions } from './handlers/generate.js';
 export type { RespectOptions } from './handlers/run.js';
-export type { JsonLogs } from './types.js';
+export * from './types.js';
+export { maskSecrets } from './modules/cli-output/mask-secrets.js';
+export { calculateTotals } from './modules/cli-output/calculate-tests-passed.js';
